+++ integration_wsl_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/firstmeet/ovnflow/internal/ovsdbjson"
+	"github.com/firstmeet/ovnflow/v2/internal/ovsdbjson"
 )
 
 const (